Solved day 8: part 1 AND part 2 (part 2 needs to be optimized, takes ~10s to run)

+++ Day07/part2.cpp
@@ -40,7 +40,7 @@ int main() {
 
 	long long int result = 0;
 	for (const long long int& t : tachyon) {
-		result += (long long int)t;
+		result += t;
 	}
